Finished review of thesis.
Left comments throughout the thesis.  In addition to the current graphs,
it would also be nice to see a summary set of graphs that show runtimes
against number of threads for the various configurations.  Overall a
nice thesis.  I especially like the suggestions for future reserach.
Definitely some important suggestions to think about.